requirements.txt
pylint==2.13.5
pytest==7.1.1
black==22.3.0
pyarrow
pandas
db-dtypes
Flask
gunicorn
google-cloud-bigquery
google-cloud-bigquery-storage